third_party/WebKit/LayoutTests/fast/events/composition-event-source-device-event-sender.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<script src="../../resources/js-test.js"></script>
<input id="input" type="input">
<script type="text/javascript">
description('Test that composition events should have a sourceDevice and its firesTouchEvents should always be false.');

function compositionHandler() {
    debug(event.type);
    shouldBeNonNull("event.sourceDevice");
    shouldBeFalse("event.sourceDevice.firesTouchEvents");
}

if (window.eventSender) {
    for (var evt of ['compositionstart', 'compositionupdate', 'compositionend']) {
        document.addEventListener(evt, compositionHandler);
    }
    
    var input = document.getElementById('input');
    input.focus();

    textInputController.setMarkedText('abcde', 1, 3);
    
    input.blur();
} else {
   debug('This test requires eventSender.');
}

</script>
