third_party/WebKit/LayoutTests/fast/events/composition-event-source-device-event-sender-expected.txt

Test that composition events should have a sourceDevice and its firesTouchEvents should always be false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


compositionstart
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
compositionupdate
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
compositionend
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
PASS successfullyParsed is true

TEST COMPLETE
